Add common interface within admin purchase to new file
