+++ client/src/components/Ticketing/ticketingmanager/AdminPurchase/AdminPurchase.tsx
@@ -19,50 +19,7 @@ import {
   getTicketRestriction,
 } from './utils/adminApiRequests';
 import {useFetchToken} from '../showings/ShowingUpdated/ShowingUtils';
-import {current} from '@reduxjs/toolkit';
-
-interface ticketTypeRestriction {
-  concessionprice: string;
-  eventinstanceid_fk: number;
-  price: string;
-  seasontickettypepricedefaultid_fk: number | null;
-  ticketlimit: number;
-  ticketrestrictionsid: number;
-  ticketssold: number;
-  tickettypedescription?: string;
-  tickettypeid_fk: number;
-}
-
-const initialTicketTypeRestriction: ticketTypeRestriction = {
-  concessionprice: '',
-  eventinstanceid_fk: 0,
-  price: '',
-  seasontickettypepricedefaultid_fk: 0,
-  ticketlimit: 0,
-  ticketrestrictionsid: 0,
-  ticketssold: 0,
-  tickettypedescription: '',
-  tickettypeid_fk: 0,
-};
-
-export type EventRow = {
-  id?: number;
-  desc: string;
-  eventid?: number;
-  eventinstanceid?: number;
-  eventname?: string;
-  eventdate?: string;
-  eventtime?: string;
-  ticketTypes?: string;
-  ticketRestrictionInfo?: ticketTypeRestriction[];
-  price?: number;
-  complimentary?: boolean;
-  availableSeats?: number;
-  seatsForType?: number;
-  imageurl?: string;
-  qty?: number;
-  typeID?: number;
-};
+import {initialTicketTypeRestriction, EventRow} from './utils/adminCommon';
 
 const AdminPurchase = () => {
   const emptyRows: EventRow[] = [
@@ -568,11 +525,6 @@ const AdminPurchase = () => {
       void fetchAllTicketTypes();
     }
   }, [token]);
-
-  // DELETE LATER
-  useEffect(() => {
-    console.log(eventData);
-  }, [eventData]);
 
   return (
     <div className='w-full h-screen absolute'>
